Fix formatting of conditional email body text
If the user sets the target branch to master and considers the deploy
request a non-special one, then this would result in two empty lines
after the line with the name of the current branch.

If none of the two trigger, then there should only be one blank line
after the branch name.