Escape all mailto-link components
So far I've been escaping selectively with URI.escape in the "body" of
the mailto-link. This has lead to problems with the `open` command on
OSX when the "to" or "subject" needed to be escaped.

Another problem currently present is that non-escaped double quotes and
ampersands could lead to the `open` command interpreting the mailto-link
as multiple arguments, instead of one string argument. This would lead
to either the mail client being opened with only a subset of the text,
or the command not running at all.

By escaping all components at all times, both the `open` command and
`xdg-open` (with Thunderbird set as the mail client) open the
mailto-link properly.

About the usage of CGI.escape instead of URI.escape:

URI.escape doesn't escape ampersands since they're used as parameter
separators. Having non-escaped ampersands left in the text results in
the `open` command interpreting everything after the non-escaped
ampersand as another command. CGI.escape seems to do the escaping in
such a way that both `open` and `xdg-open` do the right thing, except
that it replaces spaces with pluses. A simple `gsub` fixes this.